refactor: move sidebar input support
