+++ Sources/Sidebar/SidebarInputSupport.swift
@@ -1,0 +1,44 @@
+import AppKit
+import SwiftUI
+
+struct MiddleClickCapture: NSViewRepresentable {
+    let onMiddleClick: () -> Void
+
+    func makeNSView(context: Context) -> MiddleClickCaptureView {
+        let view = MiddleClickCaptureView()
+        view.onMiddleClick = onMiddleClick
+        return view
+    }
+
+    func updateNSView(_ nsView: MiddleClickCaptureView, context: Context) {
+        nsView.onMiddleClick = onMiddleClick
+    }
+}
+
+final class MiddleClickCaptureView: NSView {
+    var onMiddleClick: (() -> Void)?
+
+    override func hitTest(_ point: NSPoint) -> NSView? {
+        // Only intercept middle-click so left-click selection and right-click context menus
+        // continue to hit-test through to SwiftUI/AppKit normally.
+        guard let event = NSApp.currentEvent,
+              event.type == .otherMouseDown,
+              event.buttonNumber == 2 else {
+            return nil
+        }
+        return self
+    }
+
+    override func otherMouseDown(with event: NSEvent) {
+        guard event.buttonNumber == 2 else {
+            super.otherMouseDown(with: event)
+            return
+        }
+        onMiddleClick?()
+    }
+}
+
+enum SidebarSelection {
+    case tabs
+    case notifications
+}
